+++ dms-backend/app/core/config.py
@@ -6,12 +6,12 @@ load_dotenv()
 
 
 class Settings:
-    PROJECT_NAME: str = "Algoholic 🔥"
+    PROJECT_NAME: str = "DMS"
     PROJECT_VERSION: str = "1.0.0"
 
     DATABASE_URL = os.getenv("DATABASE_URL")
 
-    SECRET_KEY: str = os.getenv("SECRET_KEY")
+    SECRET_KEY: str = os.getenv("SECRET_KEY", "DEFAULT_KEY")
     ALGORITHM = "HS256"
     ACCESS_TOKEN_EXPIRE_MINUTES = 30
