+++ dms-backend/app/core/security.py
@@ -1,8 +1,7 @@
 from datetime import datetime
 from datetime import timedelta
 from typing import Optional
-
-from core.config import settings
+from app.core.config import settings
 from jose import jwt
