+++ dms-backend/app/db/base.py
@@ -1,2 +1,2 @@
-from db.base_class import Base
-from db.models.user import User
+from app.db.base_class import Base
+from app.db.models.user import User
