+++ dms-backend/app/db/models/user.py
@@ -1,4 +1,4 @@
-from db.base_class import Base
+from app.db.base_class import Base
 from sqlalchemy import Boolean
 from sqlalchemy import Column
 from sqlalchemy import Integer
