+++ dms-backend/app/db/repository/login.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
-from db.models.user import User 
+from app.db.models.user import User
 
 
-def get_user(username:str,db: Session):
+def get_user(username: str, db: Session):
     user = db.query(User).filter(User.username == username).first()
     return user
