+++ dms-backend/app/db/session.py
@@ -1,6 +1,5 @@
 from typing import Generator
-
-from core.config import settings
+from app.core.config import settings
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
