+++ dms-backend/main.py
@@ -1,0 +1,52 @@
+from app.apis.base import api_router
+from app.core.config import settings
+from app.db.base import Base
+from app.db.session import engine
+from fastapi import FastAPI, Request, Response
+from fastapi.staticfiles import StaticFiles
+from fastapi.middleware.cors import CORSMiddleware
+from starlette.middleware.base import BaseHTTPMiddleware
+from starlette import status
+
+
+class CORSPreflightMiddleware(BaseHTTPMiddleware):
+    async def dispatch(self, request: Request, call_next):
+        if request.method == "OPTIONS":
+            response = Response(status_code=status.HTTP_204_NO_CONTENT)
+            response.headers["Access-Control-Allow-Origin"] = "*"
+            response.headers["Access-Control-Allow-Methods"] = "*"
+            response.headers["Access-Control-Allow-Headers"] = "*"
+            return response
+        return await call_next(request)
+
+
+def create_tables():
+    Base.metadata.create_all(bind=engine)
+
+
+def include_router(app):
+    app.include_router(api_router)
+
+
+def configure_staticfiles(app):
+    app.mount("/static", StaticFiles(directory="static"), name="static")
+
+
+def start_application():
+    app = FastAPI(title=settings.PROJECT_NAME,
+                  version=settings.PROJECT_VERSION)
+    create_tables()
+    include_router(app)
+    # configure_staticfiles(app)
+    app.add_middleware(CORSPreflightMiddleware)
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["*"],
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+    return app
+
+
+app = start_application()
