+++ dms-backend/requirements.txt
@@ -1,22 +1,33 @@
+alembic==1.12.1
 annotated-types==0.6.0
 anyio==3.7.1
+bcrypt==4.0.1
 click==8.1.7
+dnspython==2.4.2
+ecdsa==0.18.0
+email-validator==2.1.0.post1
 exceptiongroup==1.1.3
 fastapi==0.104.1
 greenlet==3.0.1
 h11==0.14.0
 idna==3.4
+jose==1.0.0
+Mako==1.3.0
 mangum==0.17.0
+MarkupSafe==2.1.3
+passlib==1.7.4
+psycopg2==2.9.9
 psycopg2-binary==2.9.6
+pyasn1==0.5.0
 pydantic==2.5.0
 pydantic_core==2.14.1
+python-dotenv==1.0.0
+python-jose==3.3.0
+python-multipart==0.0.6
+rsa==4.9
+six==1.16.0
 sniffio==1.3.0
 SQLAlchemy==2.0.23
-alembic
 starlette==0.27.0
 typing_extensions==4.8.0
 uvicorn==0.24.0.post1
-python-jose==3.3.0
-python-multipart==0.0.6 
-jose
-pydantic[email]==2.5.0
